server/requirements.txt
fastapi
uvicorn
firebase-admin
google-cloud-firestore
pydantic
python-dotenv
groq
cachetools
pillow
python-multipart